[C] Initial pass DRYing layouts
