+++ packages/admin/pages/collections/[slug]/collections.tsx
@@ -1,11 +1,8 @@
 import { graphql, usePreloadedQuery, PreloadedQuery } from "react-relay";
-import { QueryTransitionWrapper } from "@wdp/lib/api/components";
 import CollectionList from "components/composed/collection/CollectionList";
-import { useRouteSlug, useBaseListQueryVars, useSearchQueryVars } from "hooks";
-import ErrorPage from "next/error";
-import { LoadingPage } from "components/atomic";
 import CollectionLayout from "components/composed/collection/CollectionLayout";
 import { AuthContextProvider } from "contexts/AuthContext";
+import Layout from "./layout";
 import type { collectionsManageSlugCollectionsPagesQuery as Query } from "__generated__/collectionsManageSlugCollectionsPagesQuery.graphql";
 import type { GetLayout } from "@wdp/lib/types/page";
 
@@ -26,38 +23,9 @@ function CollectionChildCollections({ queryRef }: Props) {
   ) : null;
 }
 
-const getLayout: GetLayout<Props> = (props) => {
-  const queryVars = useBaseListQueryVars();
-  const searchQueryVars = useSearchQueryVars();
-
-  const collectionSlug = useRouteSlug();
-  if (!collectionSlug) return <ErrorPage statusCode={404} />;
-
-  const { PageComponent, pageComponentProps } = props;
-
-  return (
-    <QueryTransitionWrapper<Query>
-      query={query}
-      variables={{
-        ...queryVars,
-        ...searchQueryVars,
-        collectionSlug,
-      }}
-      loadingFallback={<LoadingPage />}
-      refetchTags={["collections"]}
-    >
-      {({ queryRef }) =>
-        queryRef ? (
-          <PageComponent {...pageComponentProps} queryRef={queryRef} />
-        ) : (
-          <CollectionLayout>
-            <CollectionList headerStyle="secondary" hideHeader />
-          </CollectionLayout>
-        )
-      }
-    </QueryTransitionWrapper>
-  );
-};
+const getLayout: GetLayout<Props> = (props) => (
+  <Layout query={query} refetchTags={["collections"]} {...props} />
+);
 
 CollectionChildCollections.getLayout = getLayout;
