+++ packages/admin/pages/collections/[slug]/manage/layout.tsx
@@ -1,0 +1,57 @@
+import { ComponentType } from "react";
+import { PreloadedQuery, GraphQLTaggedNode } from "react-relay";
+import { OperationType } from "relay-runtime";
+import { QueryTransitionWrapper } from "@wdp/lib/api/components";
+import { useRouteSlug, useBaseListQueryVars } from "hooks";
+import ErrorPage from "next/error";
+import { LoadingPage } from "components/atomic";
+import CollectionLayout from "components/composed/collection/CollectionLayout";
+import EntityLinksList from "components/composed/links/EntityLinksList";
+
+export default function Layout<T extends OperationType>(props: Props<T>) {
+  const queryVars = useBaseListQueryVars();
+
+  const collectionSlug = useRouteSlug();
+
+  if (!collectionSlug) return <ErrorPage statusCode={404} />;
+
+  const { PageComponent, pageComponentProps, query, refetchTags } = props;
+
+  return (
+    <QueryTransitionWrapper<T>
+      query={query}
+      variables={{ ...queryVars, collectionSlug }}
+      loadingFallback={<LoadingPage />}
+      refetchTags={refetchTags}
+    >
+      {({ queryRef }) =>
+        queryRef ? (
+          <PageComponent
+            {...pageComponentProps}
+            queryRef={queryRef}
+            showSidebar
+            useRouteHeader={false}
+          />
+        ) : (
+          <CollectionLayout showSidebar useRouteHeader={false}>
+            <EntityLinksList headerStyle="secondary" hideHeader />
+          </CollectionLayout>
+        )
+      }
+    </QueryTransitionWrapper>
+  );
+}
+
+type Props<T extends OperationType> = {
+  pageComponentProps: PageProps<T>;
+  PageComponent: ComponentType<PageProps<T>>;
+  query: GraphQLTaggedNode;
+  refetchTags?: string[];
+  showLoadingCircle?: boolean;
+};
+
+type PageProps<T extends OperationType> = {
+  queryRef: PreloadedQuery<T>;
+  showSidebar: true;
+  useRouteHeader: false;
+};
