+++ packages/admin/pages/communities/[slug]/collections.tsx
@@ -1,12 +1,9 @@
 import { graphql, usePreloadedQuery, PreloadedQuery } from "react-relay";
-import { QueryTransitionWrapper } from "@wdp/lib/api/components";
 import { collectionsSlugCommunitiesPagesQuery as Query } from "__generated__/collectionsSlugCommunitiesPagesQuery.graphql";
 import CollectionList from "components/composed/collection/CollectionList";
-import { useRouteSlug, useBaseListQueryVars, useSearchQueryVars } from "hooks";
-import ErrorPage from "next/error";
-import { LoadingPage } from "components/atomic";
 import CommunityLayout from "components/composed/community/CommunityLayout";
 import { AuthContextProvider } from "contexts/AuthContext";
+import Layout from "./layout";
 import type { GetLayout } from "@wdp/lib/types/page";
 
 function CommunityChildCollections({ queryRef }: Props) {
@@ -26,38 +23,10 @@ function CommunityChildCollections({ queryRef }: Props) {
   );
 }
 
-const getLayout: GetLayout<Props> = (props) => {
-  const queryVars = useBaseListQueryVars();
-  const searchQueryVars = useSearchQueryVars();
+const getLayout: GetLayout<Props> = (props) => (
+  <Layout query={query} refetchTags={["collections"]} {...props} />
+);
 
-  const communitySlug = useRouteSlug();
-  if (!communitySlug) return <ErrorPage statusCode={404} />;
-
-  const { PageComponent, pageComponentProps } = props;
-
-  return (
-    <QueryTransitionWrapper<Query>
-      query={query}
-      variables={{
-        ...queryVars,
-        ...searchQueryVars,
-        communitySlug,
-      }}
-      loadingFallback={<LoadingPage />}
-      refetchTags={["collections"]}
-    >
-      {({ queryRef }) =>
-        queryRef ? (
-          <PageComponent {...pageComponentProps} queryRef={queryRef} />
-        ) : (
-          <CommunityLayout>
-            <CollectionList headerStyle="secondary" hideHeader />
-          </CommunityLayout>
-        )
-      }
-    </QueryTransitionWrapper>
-  );
-};
 CommunityChildCollections.getLayout = getLayout;
 
 export default CommunityChildCollections;
