+++ packages/admin/pages/communities/[slug]/layout.tsx
@@ -1,0 +1,53 @@
+import { ComponentType } from "react";
+import { PreloadedQuery, GraphQLTaggedNode } from "react-relay";
+import { OperationType } from "relay-runtime";
+import { QueryTransitionWrapper } from "@wdp/lib/api/components";
+import CollectionList from "components/composed/collection/CollectionList";
+import { useRouteSlug, useBaseListQueryVars, useSearchQueryVars } from "hooks";
+import ErrorPage from "next/error";
+import { LoadingPage } from "components/atomic";
+import CommunityLayout from "components/composed/community/CommunityLayout";
+
+export default function Layout<T extends OperationType>(props: Props<T>) {
+  const queryVars = useBaseListQueryVars();
+  const searchQueryVars = useSearchQueryVars();
+
+  const communitySlug = useRouteSlug();
+  if (!communitySlug) return <ErrorPage statusCode={404} />;
+
+  const { PageComponent, pageComponentProps, query, refetchTags } = props;
+
+  return (
+    <QueryTransitionWrapper<T>
+      query={query}
+      variables={{
+        ...queryVars,
+        ...searchQueryVars,
+        communitySlug,
+      }}
+      loadingFallback={<LoadingPage />}
+      refetchTags={refetchTags}
+    >
+      {({ queryRef }) =>
+        queryRef ? (
+          <PageComponent {...pageComponentProps} queryRef={queryRef} />
+        ) : (
+          <CommunityLayout>
+            <CollectionList headerStyle="secondary" hideHeader />
+          </CommunityLayout>
+        )
+      }
+    </QueryTransitionWrapper>
+  );
+}
+
+type Props<T extends OperationType> = {
+  pageComponentProps: PageProps<T>;
+  PageComponent: ComponentType<PageProps<T>>;
+  query: GraphQLTaggedNode;
+  refetchTags?: string[];
+};
+
+type PageProps<T extends OperationType> = {
+  queryRef: PreloadedQuery<T>;
+};
