+++ packages/admin/pages/communities/[slug]/manage/members.tsx
@@ -1,12 +1,9 @@
 import { graphql, usePreloadedQuery, PreloadedQuery } from "react-relay";
-import { QueryTransitionWrapper } from "@wdp/lib/api/components";
 import RoleAccessList from "components/composed/role/RoleAccessList";
-import ErrorPage from "next/error";
-import { useRouteSlug, useBaseListQueryVars, useSearchQueryVars } from "hooks";
 import { AuthContextProvider } from "contexts/AuthContext";
 import CommunityLayout from "components/composed/community/CommunityLayout";
-import { LoadingPage } from "components/atomic";
 import type { membersManageSlugCommunitiesPagesQuery as Query } from "@/relay/membersManageSlugCommunitiesPagesQuery.graphql";
+import Layout from "./layout";
 import type { GetLayout } from "@wdp/lib/types/page";
 
 function CommunityMembers({ queryRef, ...layoutProps }: Props) {
@@ -25,36 +22,15 @@ function CommunityMembers({ queryRef, ...layoutProps }: Props) {
     </AuthContextProvider>
   ) : null;
 }
-const getLayout: GetLayout<Props> = (props) => {
-  const queryVars = useBaseListQueryVars();
-  const searchQueryVars = useSearchQueryVars();
 
-  const communitySlug = useRouteSlug();
-
-  if (!communitySlug) return <ErrorPage statusCode={404} />;
-
-  const { PageComponent, pageComponentProps } = props;
-
-  return (
-    <QueryTransitionWrapper<Query>
-      query={query}
-      variables={{ ...queryVars, ...searchQueryVars, communitySlug }}
-      loadingFallback={<LoadingPage />}
-      refetchTags={["allAccessGrants", "assignedUsers"]}
-    >
-      {({ queryRef }) =>
-        queryRef && (
-          <PageComponent
-            {...pageComponentProps}
-            queryRef={queryRef}
-            showSidebar
-            useRouteHeader={false}
-          />
-        )
-      }
-    </QueryTransitionWrapper>
-  );
-};
+const getLayout: GetLayout<Props> = (props) => (
+  <Layout
+    query={query}
+    refetchTags={["allAccessGrants", "assignedUsers"]}
+    useRouteHeader={false}
+    {...props}
+  />
+);
 
 CommunityMembers.getLayout = getLayout;
 
@@ -63,7 +39,7 @@ export default CommunityMembers;
 type Props = {
   queryRef: PreloadedQuery<Query>;
   showSidebar: true;
-  useRouteHeader: false;
+  useRouteHeader?: boolean;
 };
 
 const query = graphql`
