+++ packages/admin/pages/contributors/[slug]/collections.tsx
@@ -1,14 +1,8 @@
 import { graphql, usePreloadedQuery, PreloadedQuery } from "react-relay";
 import ContributorLayout from "components/composed/contributor/ContributorLayout";
-import { QueryTransitionWrapper } from "@wdp/lib/api/components";
 import CollectionContributionList from "components/composed/contribution/CollectionContributionList";
-import { LoadingPage } from "components/atomic";
-import ErrorPage from "next/error";
-import { useRouteSlug, useBaseListQueryVars, useSearchQueryVars } from "hooks";
-import type {
-  ContributionOrder,
-  collectionsSlugContributorsPagesQuery as Query,
-} from "@/relay/collectionsSlugContributorsPagesQuery.graphql";
+import type { collectionsSlugContributorsPagesQuery as Query } from "@/relay/collectionsSlugContributorsPagesQuery.graphql";
+import Layout from "./layout";
 import type { GetLayout } from "@wdp/lib/types/page";
 
 function ContributorCollectionContributions({
@@ -27,35 +21,10 @@ function ContributorCollectionContributions({
   ) : null;
 }
 
-const getLayout: GetLayout<Props> = (props) => {
-  const { order, ...queryVars } = useBaseListQueryVars();
-  const contributorSlug = useRouteSlug();
-  /* eslint-disable-next-line @typescript-eslint/no-unused-vars */
-  const _searchVars = useSearchQueryVars();
+const getLayout: GetLayout<Props> = (props) => (
+  <Layout query={query} refetchTags={["contributions"]} {...props} />
+);
 
-  if (!contributorSlug) return <ErrorPage statusCode={404} />;
-
-  const { PageComponent, pageComponentProps } = props;
-
-  return (
-    <QueryTransitionWrapper<Query>
-      query={query}
-      variables={{
-        order: order as ContributionOrder,
-        ...queryVars,
-        contributorSlug,
-      }}
-      loadingFallback={<LoadingPage />}
-      refetchTags={["contributions"]}
-    >
-      {({ queryRef }) =>
-        queryRef && (
-          <PageComponent {...pageComponentProps} queryRef={queryRef} />
-        )
-      }
-    </QueryTransitionWrapper>
-  );
-};
 ContributorCollectionContributions.getLayout = getLayout;
 
 export default ContributorCollectionContributions;
