+++ packages/admin/pages/contributors/[slug]/items.tsx
@@ -1,14 +1,8 @@
 import { graphql, usePreloadedQuery, PreloadedQuery } from "react-relay";
 import ContributorLayout from "components/composed/contributor/ContributorLayout";
-import { QueryTransitionWrapper } from "@wdp/lib/api/components";
 import ItemContributionList from "components/composed/contribution/ItemContributionList";
-import { LoadingPage } from "components/atomic";
-import ErrorPage from "next/error";
-import { useRouteSlug, useBaseListQueryVars, useSearchQueryVars } from "hooks";
-import type {
-  ContributionOrder,
-  itemsSlugContributorsPagesQuery as Query,
-} from "@/relay/itemsSlugContributorsPagesQuery.graphql";
+import type { itemsSlugContributorsPagesQuery as Query } from "@/relay/itemsSlugContributorsPagesQuery.graphql";
+import Layout from "./layout";
 import type { GetLayout } from "@wdp/lib/types/page";
 
 function ContributorItemContributions({ queryRef, ...layoutProps }: Props) {
@@ -24,39 +18,10 @@ function ContributorItemContributions({ queryRef, ...layoutProps }: Props) {
   ) : null;
 }
 
-const getLayout: GetLayout<Props> = (props) => {
-  const { order, ...queryVars } = useBaseListQueryVars();
-  const contributorSlug = useRouteSlug();
-  /* eslint-disable-next-line @typescript-eslint/no-unused-vars */
-  const _searchVars = useSearchQueryVars();
+const getLayout: GetLayout<Props> = (props) => (
+  <Layout query={query} refetchTags={["contributions"]} {...props} />
+);
 
-  if (!contributorSlug) return <ErrorPage statusCode={404} />;
-
-  const { PageComponent, pageComponentProps } = props;
-
-  return (
-    <QueryTransitionWrapper<Query>
-      query={query}
-      variables={{
-        order: order as ContributionOrder,
-        ...queryVars,
-        contributorSlug,
-      }}
-      loadingFallback={<LoadingPage />}
-      refetchTags={["contributions"]}
-    >
-      {({ queryRef }) =>
-        queryRef ? (
-          <PageComponent {...pageComponentProps} queryRef={queryRef} />
-        ) : (
-          <ContributorLayout>
-            <ItemContributionList headerStyle="secondary" />
-          </ContributorLayout>
-        )
-      }
-    </QueryTransitionWrapper>
-  );
-};
 ContributorItemContributions.getLayout = getLayout;
 
 export default ContributorItemContributions;
