+++ packages/admin/pages/items/[slug]/index.tsx
@@ -1,13 +1,10 @@
 import { GetLayout } from "@wdp/lib/types/page";
 import { graphql, usePreloadedQuery, PreloadedQuery } from "react-relay";
-import { QueryTransitionWrapper } from "@wdp/lib/api/components";
 import ItemSlugRedirect from "components/composed/item/ItemSlugRedirect";
-import { useRouteSlug, useBaseListQueryVars, useSearchQueryVars } from "hooks";
-import ErrorPage from "next/error";
-import { LoadingPage } from "components/atomic";
 import { AuthContextProvider } from "contexts/AuthContext";
 import ItemLayout from "components/composed/item/ItemLayout";
 import { SlugItemsPageQuery as Query } from "@/relay/SlugItemsPageQuery.graphql";
+import Layout from "./layout";
 
 function ItemSlug({ queryRef }: Props) {
   const { item } = usePreloadedQuery<Query>(query, queryRef);
@@ -23,40 +20,12 @@ function ItemSlug({ queryRef }: Props) {
   );
 }
 
+const getLayout: GetLayout<Props> = (props) => (
+  <Layout query={query} {...props} />
+);
+
 type Props = {
   queryRef: PreloadedQuery<Query>;
-};
-
-const getLayout: GetLayout<Props> = (props) => {
-  const queryVars = useBaseListQueryVars();
-  const searchQueryVars = useSearchQueryVars();
-
-  const itemSlug = useRouteSlug();
-  if (!itemSlug) return <ErrorPage statusCode={404} />;
-
-  const { PageComponent, pageComponentProps } = props;
-
-  return (
-    <QueryTransitionWrapper<Query>
-      query={query}
-      variables={{
-        ...queryVars,
-        ...searchQueryVars,
-        itemSlug,
-      }}
-      loadingFallback={<LoadingPage />}
-    >
-      {({ queryRef }) =>
-        queryRef ? (
-          <PageComponent {...pageComponentProps} queryRef={queryRef} />
-        ) : (
-          <ItemLayout>
-            <ItemSlugRedirect />
-          </ItemLayout>
-        )
-      }
-    </QueryTransitionWrapper>
-  );
 };
 
 ItemSlug.getLayout = getLayout;
