+++ packages/admin/pages/items/[slug]/items.tsx
@@ -1,11 +1,8 @@
 import { graphql, usePreloadedQuery, PreloadedQuery } from "react-relay";
-import { QueryTransitionWrapper } from "@wdp/lib/api/components";
 import ItemLayout from "components/composed/item/ItemLayout";
 import ItemList from "components/composed/item/ItemList";
-import { useRouteSlug, useBaseListQueryVars, useSearchQueryVars } from "hooks";
-import ErrorPage from "next/error";
-import { LoadingPage } from "components/atomic";
 import { AuthContextProvider } from "contexts/AuthContext";
+import Layout from "./layout";
 import type { itemsSlugItemsPagesQuery as Query } from "__generated__/itemsSlugItemsPagesQuery.graphql";
 import type { GetLayout } from "@wdp/lib/types/page";
 
@@ -26,38 +23,9 @@ function ItemChildItems({ queryRef }: Props) {
   ) : null;
 }
 
-const getLayout: GetLayout<Props> = (props) => {
-  const queryVars = useBaseListQueryVars();
-  const searchQueryVars = useSearchQueryVars();
-
-  const itemSlug = useRouteSlug();
-  if (!itemSlug) return <ErrorPage statusCode={404} />;
-
-  const { PageComponent, pageComponentProps } = props;
-
-  return (
-    <QueryTransitionWrapper<Query>
-      query={query}
-      variables={{
-        ...queryVars,
-        ...searchQueryVars,
-        itemSlug,
-      }}
-      loadingFallback={<LoadingPage />}
-      refetchTags={["items"]}
-    >
-      {({ queryRef }) =>
-        queryRef ? (
-          <PageComponent {...pageComponentProps} queryRef={queryRef} />
-        ) : (
-          <ItemLayout>
-            <ItemList headerStyle="secondary" hideHeader />
-          </ItemLayout>
-        )
-      }
-    </QueryTransitionWrapper>
-  );
-};
+const getLayout: GetLayout<Props> = (props) => (
+  <Layout query={query} refetchTags={["items"]} {...props} />
+);
 
 ItemChildItems.getLayout = getLayout;
