+++ packages/admin/pages/users/[slug]/details.tsx
@@ -1,12 +1,8 @@
 import { graphql, usePreloadedQuery, PreloadedQuery } from "react-relay";
-import { QueryTransitionWrapper } from "@wdp/lib/api/components";
 import UserUpdateForm from "components/composed/user/UserUpdateForm";
-import { LoadingPage } from "components/atomic";
-import { useRouteSlug, useBaseListQueryVars, useSearchQueryVars } from "hooks";
 import UserLayout from "components/composed/user/UserLayout";
-import ErrorPage from "next/error";
-import { LoadingCircle } from "components/atomic";
 import type { detailsManageSlugUsersPagesQuery as Query } from "@/relay/detailsManageSlugUsersPagesQuery.graphql";
+import Layout from "./layout";
 import type { GetLayout } from "@wdp/lib/types/page";
 
 function UserDetails({ queryRef, ...layoutProps }: Props) {
@@ -19,38 +15,10 @@ function UserDetails({ queryRef, ...layoutProps }: Props) {
   ) : null;
 }
 
-const getLayout: GetLayout<Props> = (props) => {
-  const queryVars = useBaseListQueryVars();
-  const userSlug = useRouteSlug();
-  /* eslint-disable-next-line @typescript-eslint/no-unused-vars */
-  const _searchVars = useSearchQueryVars();
+const getLayout: GetLayout<Props> = (props) => (
+  <Layout query={query} {...props} />
+);
 
-  if (!userSlug) return <ErrorPage statusCode={404} />;
-
-  const { PageComponent, pageComponentProps } = props;
-
-  return (
-    <QueryTransitionWrapper<Query>
-      query={query}
-      variables={{ ...queryVars, userSlug }}
-      loadingFallback={<LoadingPage />}
-    >
-      {({ queryRef }) =>
-        queryRef ? (
-          <PageComponent
-            {...pageComponentProps}
-            queryRef={queryRef}
-            showSidebar
-          />
-        ) : (
-          <UserLayout showSidebar>
-            <LoadingCircle className="l-page-loading" />
-          </UserLayout>
-        )
-      }
-    </QueryTransitionWrapper>
-  );
-};
 UserDetails.getLayout = getLayout;
 
 export default UserDetails;
