+++ packages/admin/pages/users/[slug]/layout.tsx
@@ -1,0 +1,62 @@
+import { ComponentType } from "react";
+import { PreloadedQuery, GraphQLTaggedNode } from "react-relay";
+import { OperationType } from "relay-runtime";
+import { QueryTransitionWrapper } from "@wdp/lib/api/components";
+import { LoadingPage } from "components/atomic";
+import { useRouteSlug, useBaseListQueryVars } from "hooks";
+import UserLayout from "components/composed/user/UserLayout";
+import ErrorPage from "next/error";
+import { LoadingCircle } from "components/atomic";
+
+export default function Layout<T extends OperationType>(props: Props<T>) {
+  const queryVars = useBaseListQueryVars();
+  const userSlug = useRouteSlug();
+
+  if (!userSlug) return <ErrorPage statusCode={404} />;
+
+  const {
+    PageComponent,
+    pageComponentProps,
+    query,
+    refetchTags,
+    useRouteHeader,
+  } = props;
+
+  return (
+    <QueryTransitionWrapper<T>
+      query={query}
+      variables={{ ...queryVars, userSlug }}
+      loadingFallback={<LoadingPage />}
+      refetchTags={refetchTags}
+    >
+      {({ queryRef }) =>
+        queryRef ? (
+          <PageComponent
+            {...pageComponentProps}
+            queryRef={queryRef}
+            showSidebar
+            useRouteHeader={useRouteHeader}
+          />
+        ) : (
+          <UserLayout showSidebar>
+            <LoadingCircle className="l-page-loading" />
+          </UserLayout>
+        )
+      }
+    </QueryTransitionWrapper>
+  );
+}
+
+type Props<T extends OperationType> = {
+  pageComponentProps: PageProps<T>;
+  PageComponent: ComponentType<PageProps<T>>;
+  query: GraphQLTaggedNode;
+  refetchTags?: string[];
+  useRouteHeader?: boolean;
+};
+
+type PageProps<T extends OperationType> = {
+  queryRef: PreloadedQuery<T>;
+  showSidebar: true;
+  useRouteHeader?: boolean;
+};
